Extract Compute operation helpers to gcp_range_cell_ops (Sonar 500-line cap)
gcp_range_cells.py exceeded SonarCloud's 500-line file cap after the shared-vpc
changes. Move the self-contained Compute long-running-operation helpers
(_wait_for_operation and friends) into gcp_range_cell_ops.py; no behavior
change.
